requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
imagehash>=4.3.1
easyocr>=1.7.0
numpy>=1.24.0
tqdm>=4.66.0
openai-whisper>=20231117
anthropic>=0.34.0
python-dotenv>=1.0.0
